requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
astunparse==1.6.3
betterproto==2.0.0b6
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
dnspython==2.7.0
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.8
flatbuffers==25.1.24
fonttools==4.55.8
gast  # Se eliminó la versión fija para evitar conflictos
google-pasta==0.2.0
grpcio==1.70.0
grpclib==0.4.7
h11==0.14.0
h2==4.1.0
h5py==3.12.1
hpack==4.1.0
hyperframe==6.1.0
id==1.5.0
idna==3.10
in-toto-attestation==0.9.3
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kagglehub==0.3.7
keras  # Se eliminó la versión fija para evitar conflictos con TensorFlow
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
model-signing==0.2.0
multidict==6.1.0
namex==0.0.8
nest-asyncio==1.6.0
numpy  # Se eliminó la versión fija para evitar conflictos con TensorFlow
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==6.1.1
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
pywin32==308; sys_platform == "win32"  # Se mantiene solo para Windows
pyzmq==26.2.1
requests==2.32.3
rfc3161-client==0.1.2
rfc8785==0.1.4
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
securesystemslib==1.2.0
sigstore==3.6.1
sigstore-protobuf-specs==0.3.2
sigstore-rekor-types==0.0.18
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.45.3
statsmodels==0.14.4
tensorboard  # Se eliminó la versión fija para evitar conflictos
tensorboard-data-server==0.7.2
tensorflow  # Se eliminó la versión fija para evitar conflictos
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
tuf==5.1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
